refactor: rebrand project from ai-flora-mind to ml-production-service
- Rename package from ai_flora_mind to ml_production_service across entire codebase
- Update environment variable prefix from FLORA_* to MPS_* for clearer namespace
- Replace FLORA_CLASSIFIER_TYPE with MPS_MODEL_TYPE for model selection
- Update API metadata to reflect ML Production Service purpose
- Refactor documentation to emphasize reference implementation nature
- Update Docker, Makefile, and CI/CD configurations with new naming
- Maintain backward compatibility for iris dataset demonstration

This rebranding better reflects the project's purpose as a reference
implementation for taking ML models from research to production, rather
than being tied to a specific "flora" classification use case.

All Tests Pass: 176 unit tests passing, 97% coverage maintained
Validation Clean: Formatting, linting, and type checking verified
