+++ ml_production_service/features.py
@@ -10,8 +10,8 @@ from typing import Any, List, Tuple
 
 import numpy as np
 
-from ai_flora_mind.configs import ModelType
-from ai_flora_mind.logging import get_logger
+from ml_production_service.configs import ModelType
+from ml_production_service.logging import get_logger
 
 logger = get_logger(__name__)
