+++ ml_production_service/predictors/random_forest.py
@@ -1,8 +1,8 @@
 """
-XGBoost predictor for iris species classification.
+Random Forest predictor for iris species classification.
 
-Implements a gradient boosting predictor using pre-trained XGBoost models
-with targeted feature engineering for maximum accuracy.
+Implements a machine learning predictor using pre-trained Random Forest models
+with comprehensive feature engineering for maximum accuracy.
 """
 
 from typing import Any
@@ -10,53 +10,52 @@ from typing import Any
 import numpy as np
 from pydantic import validate_call
 
-from ai_flora_mind.configs import IrisMeasurements, ModelType
-from ai_flora_mind.features import engineer_features, get_feature_names
-from ai_flora_mind.logging import get_logger
-from ai_flora_mind.predictors.base import BasePredictor
+from ml_production_service.configs import IrisMeasurements, ModelType
+from ml_production_service.features import engineer_features, get_feature_names
+from ml_production_service.logging import get_logger
+from ml_production_service.predictors.base import BasePredictor
 
 logger = get_logger(__name__)
 
 
-class XGBoostPredictor(BasePredictor):
+class RandomForestPredictor(BasePredictor):
     """
-    XGBoost predictor for iris species classification.
+    Random Forest predictor for iris species classification.
 
-    Loads a pre-trained XGBoost model and applies targeted feature engineering
-    for maximum accuracy. Uses 9 features (4 original + 5 engineered) optimized
-    for gradient boosting performance while preventing overfitting.
+    Loads a pre-trained Random Forest model and applies comprehensive feature engineering
+    for maximum accuracy. Uses all 14 features (4 original + 10 engineered) as identified
+    in EDA analysis.
     """
 
     model_path: str
-    model: Any = None  # Will hold the loaded XGBoost model
+    model: Any = None  # Will hold the loaded sklearn model
 
-    def __init__(self, model_path: str = "registry/prd/xgboost.joblib"):
+    def __init__(self, model_path: str = "registry/prd/random_forest.joblib"):
         super().__init__(model_path=model_path)
         self.model = self._load_model(model_path)
 
         logger.info(
-            "XGBoostPredictor initialized",
+            "RandomForestPredictor initialized",
             model_path=self.model_path,
-            model_type="XGBClassifier",
+            model_type="RandomForestClassifier",
             n_estimators=getattr(self.model, "n_estimators", "unknown"),
-            max_depth=getattr(self.model, "max_depth", "unknown"),
-            features_expected=9,  # 4 original + 5 engineered
+            features_expected=14,  # 4 original + 10 engineered
         )
 
     def _prepare_features(self, measurements: IrisMeasurements) -> np.ndarray[Any, Any]:
         """
-        Applies XGBoost feature engineering to create 9-feature vector
-        (4 original + 5 high-discriminative engineered features).
+        Applies Random Forest feature engineering to create 14-feature vector
+        (4 original + 10 engineered features).
         """
         # Convert single measurement to array format
         X = measurements.to_array().reshape(1, -1)
 
-        # Apply XGBoost feature engineering (9 features)
+        # Apply Random Forest feature engineering (all 14 features)
         feature_names = get_feature_names()
-        X_engineered, feature_names_enhanced = engineer_features(X, feature_names, ModelType.XGBOOST)
+        X_engineered, feature_names_enhanced = engineer_features(X, feature_names, ModelType.RANDOM_FOREST)
 
         logger.debug(
-            "Features prepared for XGBoost",
+            "Features prepared for Random Forest",
             original_features=len(feature_names),
             engineered_features=len(feature_names_enhanced),
             shape=X_engineered.shape,
@@ -66,7 +65,7 @@ class XGBoostPredictor(BasePredictor):
 
     @validate_call
     def predict(self, measurements: IrisMeasurements) -> str:
-        """XGBoost implementation with targeted feature engineering for maximum accuracy."""
+        """Random Forest implementation with comprehensive feature engineering."""
         logger.debug(
             "Single prediction request",
             sepal_length=measurements.sepal_length,
@@ -86,10 +85,10 @@ class XGBoostPredictor(BasePredictor):
         prediction = self.SPECIES_MAP[prediction_numeric]
 
         logger.debug(
-            "XGBoost prediction completed",
+            "Random Forest prediction completed",
             prediction=prediction,
             features_used=X_features.shape[1],
-            model_performance="theoretical_maximum_98_99_percent",
+            model_confidence="available_via_predict_proba",
         )
 
         return prediction
