+++ ml_production_service/predictors/xgboost.py
@@ -1,8 +1,8 @@
 """
-Decision Tree predictor for iris species classification.
+XGBoost predictor for iris species classification.
 
-Implements a machine learning predictor using pre-trained Decision Tree models
-with minimal feature engineering for maximum interpretability.
+Implements a gradient boosting predictor using pre-trained XGBoost models
+with targeted feature engineering for maximum accuracy.
 """
 
 from typing import Any
@@ -10,52 +10,53 @@ from typing import Any
 import numpy as np
 from pydantic import validate_call
 
-from ai_flora_mind.configs import IrisMeasurements, ModelType
-from ai_flora_mind.features import engineer_features, get_feature_names
-from ai_flora_mind.logging import get_logger
-from ai_flora_mind.predictors.base import BasePredictor
+from ml_production_service.configs import IrisMeasurements, ModelType
+from ml_production_service.features import engineer_features, get_feature_names
+from ml_production_service.logging import get_logger
+from ml_production_service.predictors.base import BasePredictor
 
 logger = get_logger(__name__)
 
 
-class DecisionTreePredictor(BasePredictor):
+class XGBoostPredictor(BasePredictor):
     """
-    Decision Tree predictor for iris species classification.
+    XGBoost predictor for iris species classification.
 
-    Loads a pre-trained Decision Tree model and applies minimal feature engineering
-    for maximum interpretability. Uses 5 features (4 original + petal_area) optimized
-    for decision tree performance.
+    Loads a pre-trained XGBoost model and applies targeted feature engineering
+    for maximum accuracy. Uses 9 features (4 original + 5 engineered) optimized
+    for gradient boosting performance while preventing overfitting.
     """
 
     model_path: str
-    model: Any = None  # Will hold the loaded sklearn model
+    model: Any = None  # Will hold the loaded XGBoost model
 
-    def __init__(self, model_path: str = "registry/prd/decision_tree.joblib"):
+    def __init__(self, model_path: str = "registry/prd/xgboost.joblib"):
         super().__init__(model_path=model_path)
         self.model = self._load_model(model_path)
 
         logger.info(
-            "DecisionTreePredictor initialized",
+            "XGBoostPredictor initialized",
             model_path=self.model_path,
-            model_type="DecisionTreeClassifier",
+            model_type="XGBClassifier",
+            n_estimators=getattr(self.model, "n_estimators", "unknown"),
             max_depth=getattr(self.model, "max_depth", "unknown"),
-            features_expected=5,  # 4 original + petal_area
+            features_expected=9,  # 4 original + 5 engineered
         )
 
     def _prepare_features(self, measurements: IrisMeasurements) -> np.ndarray[Any, Any]:
         """
-        Applies Decision Tree feature engineering to create 5-feature vector
-        (4 original + petal_area).
+        Applies XGBoost feature engineering to create 9-feature vector
+        (4 original + 5 high-discriminative engineered features).
         """
         # Convert single measurement to array format
         X = measurements.to_array().reshape(1, -1)
 
-        # Apply Decision Tree feature engineering (5 features)
+        # Apply XGBoost feature engineering (9 features)
         feature_names = get_feature_names()
-        X_engineered, feature_names_enhanced = engineer_features(X, feature_names, ModelType.DECISION_TREE)
+        X_engineered, feature_names_enhanced = engineer_features(X, feature_names, ModelType.XGBOOST)
 
         logger.debug(
-            "Features prepared for Decision Tree",
+            "Features prepared for XGBoost",
             original_features=len(feature_names),
             engineered_features=len(feature_names_enhanced),
             shape=X_engineered.shape,
@@ -65,7 +66,7 @@ class DecisionTreePredictor(BasePredictor):
 
     @validate_call
     def predict(self, measurements: IrisMeasurements) -> str:
-        """Decision Tree implementation with minimal feature engineering for interpretability."""
+        """XGBoost implementation with targeted feature engineering for maximum accuracy."""
         logger.debug(
             "Single prediction request",
             sepal_length=measurements.sepal_length,
@@ -85,10 +86,10 @@ class DecisionTreePredictor(BasePredictor):
         prediction = self.SPECIES_MAP[prediction_numeric]
 
         logger.debug(
-            "Decision Tree prediction completed",
+            "XGBoost prediction completed",
             prediction=prediction,
             features_used=X_features.shape[1],
-            model_interpretability="high_with_decision_paths",
+            model_performance="theoretical_maximum_98_99_percent",
         )
 
         return prediction
