+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
-name = "ai-flora-mind"
+name = "ml-production-service"
 version = "0.1.0"
-description = "AI Flora Mind - Python service with AI capabilities"
+description = "Reference implementation for taking ML models from research to production"
 authors = [{name="Leopoldo Garcia Vargas", email="lk13.dev@gmail.com"}]
 license = {text = "Apache-2.0"}
 requires-python = ">=3.12"
