+++ research/data.py
@@ -10,7 +10,7 @@ import numpy as np
 from sklearn.datasets import load_iris
 from sklearn.utils import Bunch
 
-from ai_flora_mind.logging import get_logger
+from ml_production_service.logging import get_logger
 
 logger = get_logger(__name__)
