+++ research/experiments/random_forest/regularized.py
@@ -19,7 +19,7 @@ import joblib
 import numpy as np
 from sklearn.ensemble import RandomForestClassifier
 
-from ai_flora_mind.logging import get_logger
+from ml_production_service.logging import get_logger
 from research.data import load_iris_data
 from research.evaluation import evaluate_model, log_performance_summary
 from research.experiments.constants import MODELS_DIR, RESULTS_DIR
