+++ tests/conftest.py
@@ -8,8 +8,13 @@ from typing import List, Tuple
 
 import pytest
 
-from ai_flora_mind.configs import IrisMeasurements
-from ai_flora_mind.predictors import DecisionTreePredictor, HeuristicPredictor, RandomForestPredictor, XGBoostPredictor
+from ml_production_service.configs import IrisMeasurements
+from ml_production_service.predictors import (
+    DecisionTreePredictor,
+    HeuristicPredictor,
+    RandomForestPredictor,
+    XGBoostPredictor,
+)
 
 
 @pytest.fixture
