+++ tests/test_configs.py
@@ -6,7 +6,7 @@ Tests the ServiceConfig class, environment variable handling, and model path res
 
 import pytest
 
-from ai_flora_mind.configs import ModelType, ServiceConfig
+from ml_production_service.configs import ModelType, ServiceConfig
 
 
 @pytest.mark.unit
@@ -17,22 +17,22 @@ def test__service_config__default_model_type_is_heuristic() -> None:
 
 @pytest.mark.unit
 def test__service_config__environment_variable_override(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setenv("FLORA_CLASSIFIER_TYPE", "random_forest")
+    monkeypatch.setenv("MPS_MODEL_TYPE", "random_forest")
     config = ServiceConfig()
     assert config.model_type == ModelType.RANDOM_FOREST
 
 
 @pytest.mark.unit
 def test__service_config__case_insensitive_environment_variables(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setenv("flora_classifier_type", "random_forest")
+    monkeypatch.setenv("mps_model_type", "random_forest")
     config = ServiceConfig()
     assert config.model_type == ModelType.RANDOM_FOREST
 
 
 @pytest.mark.unit
 def test__service_config__extra_environment_variables_ignored(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setenv("FLORA_CLASSIFIER_TYPE", "heuristic")
-    monkeypatch.setenv("FLORA_UNKNOWN_VAR", "should_be_ignored")
+    monkeypatch.setenv("MPS_MODEL_TYPE", "heuristic")
+    monkeypatch.setenv("MPS_UNKNOWN_VAR", "should_be_ignored")
     config = ServiceConfig()
     assert config.model_type == ModelType.HEURISTIC
 
@@ -45,7 +45,7 @@ def test__service_config__get_model_path_heuristic_returns_none() -> None:
 
 @pytest.mark.unit
 def test__service_config__get_model_path_random_forest_returns_consistent_path(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setenv("FLORA_CLASSIFIER_TYPE", "random_forest")
+    monkeypatch.setenv("MPS_MODEL_TYPE", "random_forest")
     config = ServiceConfig()
 
     model_path = config.get_model_path()
@@ -54,7 +54,7 @@ def test__service_config__get_model_path_random_forest_returns_consistent_path(m
 
 @pytest.mark.unit
 def test__service_config__environment_prefix_handling(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setenv("FLORA_CLASSIFIER_TYPE", "random_forest")
+    monkeypatch.setenv("MPS_MODEL_TYPE", "random_forest")
     monkeypatch.setenv("OTHER_CLASSIFIER_TYPE", "should_be_ignored")
 
     config = ServiceConfig()
@@ -63,7 +63,7 @@ def test__service_config__environment_prefix_handling(monkeypatch: pytest.Monkey
 
 @pytest.mark.unit
 def test__service_config__get_model_path_decision_tree_returns_consistent_path(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setenv("FLORA_CLASSIFIER_TYPE", "decision_tree")
+    monkeypatch.setenv("MPS_MODEL_TYPE", "decision_tree")
     config = ServiceConfig()
 
     model_path = config.get_model_path()
@@ -72,7 +72,7 @@ def test__service_config__get_model_path_decision_tree_returns_consistent_path(m
 
 @pytest.mark.unit
 def test__service_config__xgboost_model_path(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setenv("FLORA_CLASSIFIER_TYPE", "xgboost")
+    monkeypatch.setenv("MPS_MODEL_TYPE", "xgboost")
     config = ServiceConfig()
 
     model_path = config.get_model_path()
