+++ tests/test_factory.py
@@ -12,9 +12,9 @@ import joblib
 import pytest
 from sklearn.ensemble import RandomForestClassifier
 
-from ai_flora_mind.configs import ModelType, ServiceConfig
-from ai_flora_mind.factory import get_predictor
-from ai_flora_mind.predictors import (
+from ml_production_service.configs import ModelType, ServiceConfig
+from ml_production_service.factory import get_predictor
+from ml_production_service.predictors import (
     BasePredictor,
     DecisionTreePredictor,
     HeuristicPredictor,
@@ -39,14 +39,14 @@ def test__get_predictor__xgboost_model_creates_xgboost_predictor() -> None:
         pytest.skip("XGBoost model not available for testing")
 
     try:
-        os.environ["FLORA_CLASSIFIER_TYPE"] = "xgboost"
+        os.environ["MPS_MODEL_TYPE"] = "xgboost"
         config = ServiceConfig()
         predictor = get_predictor(config)
 
         assert isinstance(predictor, XGBoostPredictor)
         assert isinstance(predictor, BasePredictor)
     finally:
-        os.environ.pop("FLORA_CLASSIFIER_TYPE", None)
+        os.environ.pop("MPS_MODEL_TYPE", None)
 
 
 @pytest.mark.unit
@@ -69,7 +69,7 @@ def test__get_predictor__decision_tree_model_creates_decision_tree_predictor() -
     if not os.path.exists("registry/prd/decision_tree.joblib"):
         pytest.skip("Production decision tree model not available")
 
-    os.environ["FLORA_CLASSIFIER_TYPE"] = "decision_tree"
+    os.environ["MPS_MODEL_TYPE"] = "decision_tree"
     try:
         config = ServiceConfig()
         predictor = get_predictor(config)
@@ -77,7 +77,7 @@ def test__get_predictor__decision_tree_model_creates_decision_tree_predictor() -
         assert isinstance(predictor, DecisionTreePredictor)
         assert isinstance(predictor, BasePredictor)
     finally:
-        os.environ.pop("FLORA_CLASSIFIER_TYPE", None)
+        os.environ.pop("MPS_MODEL_TYPE", None)
 
 
 @pytest.mark.unit
@@ -85,26 +85,26 @@ def test__get_predictor__all_models_work_correctly(monkeypatch: pytest.MonkeyPat
     # Test that all models create correct predictor types
 
     # Heuristic - always works
-    monkeypatch.setenv("FLORA_CLASSIFIER_TYPE", "heuristic")
+    monkeypatch.setenv("MPS_MODEL_TYPE", "heuristic")
     config_heuristic = ServiceConfig()
     predictor_heuristic = get_predictor(config_heuristic)
     assert isinstance(predictor_heuristic, HeuristicPredictor)
 
     # Only test models if their files exist
     if os.path.exists("registry/prd/random_forest.joblib"):
-        monkeypatch.setenv("FLORA_CLASSIFIER_TYPE", "random_forest")
+        monkeypatch.setenv("MPS_MODEL_TYPE", "random_forest")
         config_rf = ServiceConfig()
         predictor_rf = get_predictor(config_rf)
         assert isinstance(predictor_rf, RandomForestPredictor)
 
     if os.path.exists("registry/prd/decision_tree.joblib"):
-        monkeypatch.setenv("FLORA_CLASSIFIER_TYPE", "decision_tree")
+        monkeypatch.setenv("MPS_MODEL_TYPE", "decision_tree")
         config_dt = ServiceConfig()
         predictor_dt = get_predictor(config_dt)
         assert isinstance(predictor_dt, DecisionTreePredictor)
 
     if os.path.exists("registry/prd/xgboost.joblib"):
-        monkeypatch.setenv("FLORA_CLASSIFIER_TYPE", "xgboost")
+        monkeypatch.setenv("MPS_MODEL_TYPE", "xgboost")
         config_xgb = ServiceConfig()
         predictor_xgb = get_predictor(config_xgb)
         assert isinstance(predictor_xgb, XGBoostPredictor)
@@ -115,7 +115,7 @@ def test__get_predictor__random_forest_model_creates_random_forest_predictor() -
     if not os.path.exists("registry/prd/random_forest.joblib"):
         pytest.skip("Production random forest model not available")
 
-    os.environ["FLORA_CLASSIFIER_TYPE"] = "random_forest"
+    os.environ["MPS_MODEL_TYPE"] = "random_forest"
     try:
         config = ServiceConfig()
         predictor = get_predictor(config)
@@ -123,7 +123,7 @@ def test__get_predictor__random_forest_model_creates_random_forest_predictor() -
         assert isinstance(predictor, RandomForestPredictor)
         assert isinstance(predictor, BasePredictor)
     finally:
-        os.environ.pop("FLORA_CLASSIFIER_TYPE", None)
+        os.environ.pop("MPS_MODEL_TYPE", None)
 
 
 @pytest.mark.unit
@@ -142,7 +142,7 @@ def test__get_predictor__random_forest_with_temporary_model() -> None:
 
 @pytest.mark.unit
 def test__get_predictor__random_forest_with_invalid_path_raises_error(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setenv("FLORA_CLASSIFIER_TYPE", "random_forest")
+    monkeypatch.setenv("MPS_MODEL_TYPE", "random_forest")
 
     def mock_get_model_path(self) -> str:
         return "/nonexistent/path/model.joblib"
@@ -165,7 +165,7 @@ def test__get_predictor__environment_integration_with_is_container(monkeypatch: 
         try:
             # Test in container environment
             monkeypatch.setenv("IS_CONTAINER", "true")
-            monkeypatch.setenv("FLORA_CLASSIFIER_TYPE", "random_forest")
+            monkeypatch.setenv("MPS_MODEL_TYPE", "random_forest")
 
             # Mock get_model_path to simulate container path
             def mock_get_model_path(self) -> str:
@@ -204,7 +204,7 @@ def test__get_predictor__random_forest_model_attributes_logging(caplog: pytest.L
         pytest.skip("Production random forest model not available")
 
     # Use environment variable to set model type
-    os.environ["FLORA_CLASSIFIER_TYPE"] = "random_forest"
+    os.environ["MPS_MODEL_TYPE"] = "random_forest"
     try:
         config = ServiceConfig()
 
@@ -222,12 +222,12 @@ def test__get_predictor__random_forest_model_attributes_logging(caplog: pytest.L
         assert len(rf_success_logs) > 0
     finally:
         # Clean up environment
-        os.environ.pop("FLORA_CLASSIFIER_TYPE", None)
+        os.environ.pop("MPS_MODEL_TYPE", None)
 
 
 @pytest.mark.unit
 def test__get_predictor__value_error_from_random_forest_creation(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setenv("FLORA_CLASSIFIER_TYPE", "random_forest")
+    monkeypatch.setenv("MPS_MODEL_TYPE", "random_forest")
     config = ServiceConfig()
 
     # Mock get_model_path to return None to trigger ValueError
@@ -242,7 +242,7 @@ def test__get_predictor__value_error_from_random_forest_creation(monkeypatch: py
 
 @pytest.mark.unit
 def test__get_predictor__value_error_from_decision_tree_creation(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setenv("FLORA_CLASSIFIER_TYPE", "decision_tree")
+    monkeypatch.setenv("MPS_MODEL_TYPE", "decision_tree")
     config = ServiceConfig()
 
     # Mock get_model_path to return None to trigger ValueError
@@ -257,7 +257,7 @@ def test__get_predictor__value_error_from_decision_tree_creation(monkeypatch: py
 
 @pytest.mark.unit
 def test__get_predictor__value_error_from_xgboost_creation(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setenv("FLORA_CLASSIFIER_TYPE", "xgboost")
+    monkeypatch.setenv("MPS_MODEL_TYPE", "xgboost")
     config = ServiceConfig()
 
     # Mock get_model_path to return None to trigger ValueError
